Refactor package structure and rename classes in sadu-queries2 module
This commit restructures package locations, renames several classes and interfaces, and updates import statements in the sadu-queries2 module. Many classes have moved from the "stages" and "params" packages to the "query", "parameter", "api.base" and "api.parameter" packages, respecting a semantic organization. Moreover, some classes and interfaces have been renamed to harmonize the naming scheme, improving readability and comprehension of the codebase.
