+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/api/base/ConnectionProvider.java
@@ -4,7 +4,7 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 
-package de.chojo.sadu.queries.stages.base;
+package de.chojo.sadu.queries.api.base;
 
 import de.chojo.sadu.base.DataSourceProvider;
 import de.chojo.sadu.exceptions.ThrowingFunction;
