+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/api/parameter/BaseParameter.java
@@ -4,7 +4,7 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 
-package de.chojo.sadu.queries.params;
+package de.chojo.sadu.queries.api.parameter;
 
 import de.chojo.sadu.queries.query.TokenizedQuery;
 
@@ -12,6 +12,6 @@ import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 @SuppressWarnings("InterfaceMayBeAnnotatedFunctional")
-public interface BaseParam {
+public interface BaseParameter {
     void apply(TokenizedQuery query, PreparedStatement stmt) throws SQLException;
 }
