+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/api/query/Query.java
@@ -7,8 +7,8 @@
 package de.chojo.sadu.queries.api.query;
 
 import de.chojo.sadu.queries.configuration.QueryConfiguration;
-import de.chojo.sadu.queries.stages.ParsedQueryImpl;
-import de.chojo.sadu.queries.stages.QueryImpl;
+import de.chojo.sadu.queries.query.ParsedQueryImpl;
+import de.chojo.sadu.queries.query.QueryImpl;
 import org.intellij.lang.annotations.Language;
 
 /**
