+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/calls/CallSupplier.java
@@ -4,10 +4,10 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 
-package de.chojo.sadu.queries.stages.calls;
+package de.chojo.sadu.queries.calls;
 
 import de.chojo.sadu.queries.api.call.calls.Calls;
-import de.chojo.sadu.queries.storage.ResultStorage;
+import de.chojo.sadu.queries.api.storage.ResultStorage;
 
 @FunctionalInterface
 public interface CallSupplier<T extends Calls> {
