+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/execution/reading/ReaderImpl.java
@@ -8,15 +8,15 @@ package de.chojo.sadu.queries.stages.execution.reading;
 
 import de.chojo.sadu.mapper.MapperConfig;
 import de.chojo.sadu.mapper.rowmapper.RowMapping;
+import de.chojo.sadu.queries.api.base.QueryProvider;
 import de.chojo.sadu.queries.api.call.Call;
 import de.chojo.sadu.queries.api.execution.reading.Reader;
 import de.chojo.sadu.queries.api.query.AppendedQuery;
 import de.chojo.sadu.queries.api.results.reading.Result;
 import de.chojo.sadu.queries.call.CallImpl;
+import de.chojo.sadu.queries.query.AppendedQueryImpl;
+import de.chojo.sadu.queries.query.QueryImpl;
 import de.chojo.sadu.queries.query.TokenizedQuery;
-import de.chojo.sadu.queries.stages.AppendedQueryImpl;
-import de.chojo.sadu.queries.stages.QueryImpl;
-import de.chojo.sadu.queries.stages.base.QueryProvider;
 import de.chojo.sadu.queries.stages.execution.writing.CalledSingletonQueryImpl;
 import de.chojo.sadu.queries.stages.results.reading.MultiResult;
 import de.chojo.sadu.queries.stages.results.reading.SingleResult;
