+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/execution/writing/CalledBatchQueryImpl.java
@@ -6,14 +6,14 @@
 
 package de.chojo.sadu.queries.stages.execution.writing;
 
+import de.chojo.sadu.queries.api.base.QueryProvider;
 import de.chojo.sadu.queries.api.execution.writing.CalledBatchQuery;
 import de.chojo.sadu.queries.api.results.writing.ManipulationBatchResult;
 import de.chojo.sadu.queries.api.results.writing.ManipulationResult;
 import de.chojo.sadu.queries.call.CallImpl;
 import de.chojo.sadu.queries.calls.BatchCall;
-import de.chojo.sadu.queries.stages.ParsedQueryImpl;
-import de.chojo.sadu.queries.stages.QueryImpl;
-import de.chojo.sadu.queries.stages.base.QueryProvider;
+import de.chojo.sadu.queries.query.ParsedQueryImpl;
+import de.chojo.sadu.queries.query.QueryImpl;
 import de.chojo.sadu.queries.stages.results.writing.ManipulationBatchQuery;
 import de.chojo.sadu.queries.stages.results.writing.ManipulationResultImpl;
