+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/execution/writing/CalledSingletonQueryImpl.java
@@ -8,16 +8,16 @@ package de.chojo.sadu.queries.stages.execution.writing;
 
 import de.chojo.sadu.mapper.MapperConfig;
 import de.chojo.sadu.mapper.rowmapper.RowMapping;
+import de.chojo.sadu.queries.api.base.QueryProvider;
 import de.chojo.sadu.queries.api.call.Call;
 import de.chojo.sadu.queries.api.execution.reading.Reader;
 import de.chojo.sadu.queries.api.execution.writing.CalledSingletonQuery;
 import de.chojo.sadu.queries.api.results.writing.ManipulationResult;
 import de.chojo.sadu.queries.call.CallImpl;
 import de.chojo.sadu.queries.calls.SingletonCall;
+import de.chojo.sadu.queries.query.ParsedQueryImpl;
+import de.chojo.sadu.queries.query.QueryImpl;
 import de.chojo.sadu.queries.query.TokenizedQuery;
-import de.chojo.sadu.queries.stages.ParsedQueryImpl;
-import de.chojo.sadu.queries.stages.QueryImpl;
-import de.chojo.sadu.queries.stages.base.QueryProvider;
 import de.chojo.sadu.queries.stages.execution.reading.AutoMappedQuery;
 import de.chojo.sadu.queries.stages.execution.reading.MappedQuery;
 import de.chojo.sadu.queries.stages.results.writing.ManipulationResultImpl;
