+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/results/reading/MultiResult.java
@@ -6,9 +6,9 @@
 
 package de.chojo.sadu.queries.stages.results.reading;
 
+import de.chojo.sadu.queries.api.base.QueryProvider;
 import de.chojo.sadu.queries.api.results.reading.Result;
-import de.chojo.sadu.queries.stages.QueryImpl;
-import de.chojo.sadu.queries.stages.base.QueryProvider;
+import de.chojo.sadu.queries.query.QueryImpl;
 import de.chojo.sadu.queries.stages.execution.reading.MappedQuery;
 
 /**
