+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/storage/ResultStorageImpl.java
@@ -7,6 +7,7 @@
 package de.chojo.sadu.queries.storage;
 
 import de.chojo.sadu.queries.api.results.reading.Result;
+import de.chojo.sadu.queries.api.storage.ResultStorage;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -16,25 +17,28 @@ import java.util.Objects;
 import java.util.Optional;
 
 @SuppressWarnings("unchecked")
-public class ResultStorage {
+public class ResultStorageImpl implements ResultStorage {
     private final Map<String, Result<?>> storage = new HashMap<>();
 
-    public ResultStorage() {
+    public ResultStorageImpl() {
     }
 
     public void store(String key, Result<?> value) {
         storage.put(key, value);
     }
 
+    @Override
     @SuppressWarnings("unchecked")
     public <T> T get(String key) {
         return (T) retrieve(key).result();
     }
 
+    @Override
     public <T> Optional<T> getAs(String key, Class<T> clazz) {
         return Optional.ofNullable((T) retrieve(key).result());
     }
 
+    @Override
     public <T> List<T> getList(String key, Class<T> clazz) {
         return Objects.requireNonNullElse((List<T>) retrieve(key).result(), Collections.emptyList());
     }
